# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DOGFOOD 2026 Hackathon Platform API",
    "version": "1.0.0",
    "description": "Comprehensive REST API specification for the DOGFOOD 2026 Hackathon Platform, covering authentication, team formation, project submissions, judging, community voting, audit logs, and cryptographic attestations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local / Container development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Django session cookie authentication (e.g. Cookie: session=...)"
      }
    },
    "schemas": {
...
      "BulkImportSummary": {
        "type": "object",
        "required": ["created", "updated", "skipped", "errors"],
        "properties": {
          "created": { "type": "integer", "example": 10 },
          "updated": { "type": "integer", "example": 2 },
          "skipped": { "type": "integer", "example": 0 },
          "errors": {
            "type": "array",
            "items": { "type": "string" },
            "example": []
          }
        }
      },
      "BulkExportPayload": {
        "type": "object",
        "properties": {
          "event": { "type": "object" },
          "events": { "type": "array", "items": { "type": "object" } },
          "tracks": { "type": "array", "items": { "type": "object" } },
          "teams": { "type": "array", "items": { "type": "object" } },
          "projects": { "type": "array", "items": { "type": "object" } }
        }
      },
      "JudgeScoreItem": {
        "type": "object",
        "properties": {
          "project_id": { "type": "integer" },
          "project_title": { "type": "string" },
          "track": { "type": "string" },
          "criteria_scores": { "type": "object", "additionalProperties": { "type": "number" } },
          "comment": { "type": "string" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      }
    }
  },
  "paths": {
...
    "/api/judge/scores/": {
      "get": {
        "summary": "List judge evaluations with strict role isolation",
        "tags": ["Judging"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "name": "judge",
            "in": "query",
            "required": false,
            "description": "Target judge username or ID (peers blocked for role 'judge')",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "List of scored projects and rubrics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/JudgeScoreItem" }
                }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Unauthorized role or peer score access attempt" }
        }
      }
    },
    "/api/export.csv": {
      "get": {
        "summary": "Export project evaluations as RFC 4180 CSV",
        "tags": ["Judging"],
        "security": [{ "sessionAuth": [] }],
        "responses": {
          "200": {
            "description": "CSV attachment with headers: project_id,project_title,track,submitted_at,review_count",
            "content": { "text/csv": { "schema": { "type": "string" } } }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Only organizer or admin roles allowed" }
        }
      }
    },
...
    "/api/export/bulk/": {
      "get": {
        "summary": "Bulk JSON export of event metadata with secret scrubbing",
        "tags": ["T4 Bulk Operations"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          { "name": "event_id", "in": "query", "required": false, "schema": { "type": "integer" } }
        ],
        "responses": {
          "200": {
            "description": "Sanitized JSON data containing events, tracks, teams, and projects",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkExportPayload" }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Organizer or admin role required" }
        }
      }
    },
    "/api/import/bulk/": {
      "post": {
        "summary": "Bulk JSON import with whole-payload pre-validation, atomicity, and dry-run",
        "tags": ["T4 Bulk Operations"],
        "security": [{ "sessionAuth": [] }],
        "parameters": [
          {
            "name": "dry_run",
            "in": "query",
            "required": false,
            "description": "If true, simulates validation and returns expected changes without modifying DB",
            "schema": { "type": "boolean", "default": false }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/BulkExportPayload" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Import / dry-run summary",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkImportSummary" }
              }
            }
          },
          "400": {
            "description": "Pre-validation failed or relational constraint violated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/BulkImportSummary" }
              }
            }
          },
          "401": { "description": "Unauthenticated" },
          "403": { "description": "Organizer or admin role required" }
        }
      }
    },
...
  }
}